Moved reflection to own namespace
