+++ tests/Unit/Executor/ParameterSetExtractorTest.php
@@ -55,7 +55,7 @@ class ParameterSetExtractorTest extends TestCase
 
     private function provideParams(array $providers)
     {
-        $payload = new Payload(__DIR__ . '/../../../lib/Benchmark/Remote/template/parameter_set_extractor.template', [
+        $payload = new Payload(__DIR__ . '/../../../lib/Reflection/template/parameter_set_extractor.template', [
             'bootstrap' => __DIR__ . '/benchmarks/ParamProviderBench.php',
             'file' => __DIR__ . '/benchmarks/ParamProviderBench.php',
             'class' => ParamProviderBench::class,
